Refactor packet ID handling to use int instead of short; update related classes and methods for consistency

+++ loadbalancer/src/main/java/dev/sweety/netty/loadbalancer/packets/MetricsUpdatePacket.java
@@ -14,7 +14,7 @@ public class MetricsUpdatePacket extends Packet {
         this.buffer().writeVarInt((int) ramUsage * 100);
     }
 
-    public MetricsUpdatePacket(short _id, long _timestamp, byte[] _data) {
+    public MetricsUpdatePacket(int _id, long _timestamp, byte[] _data) {
         super(_id, _timestamp, _data);
         this.cpuLoad = this.buffer().readVarInt() * 0.01f;
         this.ramUsage = this.buffer().readVarInt() * 0.01f;
